Switch GalleryMedia takenAt to takenAtLocal
PhotoPrism TakenAt is not guaranteed to be in UTC. TakenAtLocal is
reliable and actually correct date to use.
Date formats and temporal operations with GalleryMedia.takenAtLocal are
now UTCfied and ignore device timezone. If a photo is taken on July 4th
at 23:18 in Montenegro, its local time 2023-07-04T23:18:32Z must be
displayed and treated as July 4th, 23:18 regardless of whether
the gallery is viewed from Montenegro or USA.