Refactor GallerySearchView config view management
Make the view model fully in charge of the configuration view state.
Instead of opening the configuration and notifying the VM that it is
being opened, the view now notifies the VM on corresponding clicks
(bar, back button) and the VM switches the state accordingly.
Also search state names are shortened.