Message for 2024
In 2024, starting from the very first day, me and my wife will do our
best to save our lives, our freedom and our hope for a brighter future.
Whatever happens to us is not the end, we can deal with it. We have our
friends and families to help.
DONT TREAD ON US