Fix infinite download progress on some envs
For some unknown reason, perhaps related to HTTP/2, the request body in
OkHttpObservableDownloader is not explicitly closed on some environments,
for example when connected to demo.photoprism.app.

The downloader made simple and not relying on ResponseBody::onClose call.
Instead, the completion is emmitted once everything is written to the sink.